refactor(bridge): model 패키지 분리 및 에러 처리 테스트 추가
