+++ src/main/java/model/Bridge.java
@@ -1,4 +1,4 @@
-package bridge;
+package model;
 
 import type.MovingType;
