+++ src/main/java/model/Bridges.java
@@ -1,5 +1,6 @@
-package bridge;
+package model;
 
+import model.Bridge;
 import type.MovingType;
 
 import java.util.ArrayList;
