+++ src/main/java/model/Moving.java
@@ -1,5 +1,6 @@
-package bridge;
+package model;
 
+import model.Bridge;
 import type.MovingType;
 import validation.Validation;
