+++ src/main/java/model/Restart.java
@@ -1,4 +1,4 @@
-package bridge;
+package model;
 
 import validation.Validation;
