+++ src/test/java/bridge/ApplicationTest.java
@@ -9,6 +9,7 @@ import camp.nextstep.edu.missionutils.test.NsTest;
 
 import java.util.List;
 
+import model.BridgeSize;
 import org.junit.jupiter.api.Test;
 
 class ApplicationTest extends NsTest {
@@ -66,6 +67,25 @@ class ApplicationTest extends NsTest {
         assertRandomNumberInRangeTest(() -> {
             run("3", "U", "U", "Q");
             assertThat(output()).contains(
+                    "최종 게임 결과",
+                    "[ O | X ]",
+                    "[   |   ]",
+                    "게임 성공 여부: 실패",
+                    "총 시도한 횟수: 1"
+            );
+
+            int upSideIndex = output().indexOf("[ O | X ]");
+            int downSideIndex = output().indexOf("[   |   ]");
+            assertThat(upSideIndex).isLessThan(downSideIndex);
+        }, 1, 0, 1);
+    }
+
+    @Test
+    void 잘못된_값_입력시_에러_처리_테스트() {
+        assertRandomNumberInRangeTest(() -> {
+            run("3", "U", "F", "U", "Q");
+            assertThat(output()).contains(
+                    "[ERROR] 이동할 칸은 위는 U, 아래는 D로 입력해야 합니다.",
                     "최종 게임 결과",
                     "[ O | X ]",
                     "[   |   ]",
