+++ src/test/java/bridge/BridgeGameTest.java
@@ -1,5 +1,8 @@
 package bridge;
 
+import model.Bridge;
+import model.BridgeSize;
+import model.Moving;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
