+++ src/test/java/bridge/MovingTest.java
@@ -1,5 +1,7 @@
 package bridge;
 
+import model.Bridge;
+import model.Moving;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
