+++ src/test/java/validation/ValidationTest.java
@@ -1,6 +1,6 @@
 package validation;
 
-import bridge.BridgeSize;
+import model.BridgeSize;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
